requirements.txt
supabase-py>=1.0.0
pdfplumber>=0.7.0
requests>=2.0.0
python-dotenv>=1.0.0
